requirements.txt
pandas==2.1.0
scikit-learn==1.3.0
pydantic==2.5.0
pyyaml==6.0
dvc==3.30.0
joblib==1.3.0
pytest==7.4.0
mlflow==2.9.0
fastapi==0.109.0
uvicorn==0.27.0
numpy==1.26.0
